+++ frontend/src/App.vue
@@ -1,73 +1,19 @@
 <script setup lang="ts">
-import { ref, onMounted } from 'vue';
 import TitleComponent from './components/TitleComponent.vue';
 import TaskInputComponent from './components/TaskInputComponent.vue';
 import TaskItemComponent from './components/TaskItemComponent.vue';
-import axios from 'axios';
+import { useTaskManager } from './composables/useTaskManager';
 
-interface Task {
-  id: number;
-  task: string;
-}
-
-const tasks = ref<Task[]>([]);
-const editingIndex = ref<number | null>(null);
-
-const loadTasks = async () => {
-  try {
-    const response = await axios.get<Task[]>('/task');
-    tasks.value = response.data;
-  } catch (error) {
-    console.error('タスクの取得に失敗しました:', error);
-  }
-};
-
-const addNewTask = async (task: string) => {
-  try {
-    const response = await axios.post<Task>('/task', { task });
-    tasks.value.push(response.data); // 新しいタスクをリストに追加
-  } catch (error) {
-    console.error('タスクの追加に失敗しました:', error);
-  }
-};
-
-const deleteTask = async (index: number) => {
-  try {
-    const taskToDelete = tasks.value[index];
-    await axios.delete(`/task/${taskToDelete.id}`); // タスクIDで削除
-    tasks.value.splice(index, 1); // UI上でもタスクを削除
-    if (editingIndex.value !== null && editingIndex.value > index) {
-      editingIndex.value--;
-    } else if (editingIndex.value === index) {
-      editingIndex.value = null;
-    }
-  } catch (error) {
-    console.error('タスクの削除に失敗しました:', error);
-  }
-};
-
-const saveTask = async (index: number, task: string) => {
-  try {
-    const taskToUpdate = tasks.value[index];
-    await axios.patch(`/task/${taskToUpdate.id}`, { task });
-    tasks.value[index].task = task; // UI上でタスクを更新
-    editingIndex.value = null;
-  } catch (error) {
-    console.error('タスクの更新に失敗しました:', error);
-  }
-};
-
-const startEditing = (index: number) => {
-  editingIndex.value = index;
-};
-
-const cancelEditing = () => {
-  editingIndex.value = null;
-};
-
-onMounted(() => {
-  loadTasks();
-});
+const {
+  tasks,
+  editingIndex,
+  loadTasks,
+  addNewTask,
+  deleteTask,
+  saveTask,
+  startEditing,
+  cancelEditing
+} = useTaskManager();
 </script>
 
 
@@ -78,9 +24,9 @@ onMounted(() => {
     <TaskInputComponent :addButtonPushed="addNewTask" @onNewTaskFocus="cancelEditing"/>
     <ul>
       <TaskItemComponent
-        v-for="(task, index) in tasks"
+        v-for="(taskData, index) in tasks"
         :key="index"
-        :task="task.task"
+        :task="taskData.task"
         :index="index"
         :isEditing="editingIndex === index"
         @deleteTask="deleteTask"
